Generalized copy url context menu, added it to team buttons
